seo/backend/requirements.txt
flask
flask-cors
flask-jwt-extended
flask-bcrypt
flask-sqlalchemy
python-dotenv

numpy
pandas
reportlab
matplotlib
requests
beautifulsoup4
textstat
openpyxl
scikit-learn
textblob
google-api-python-client
instaloader
youtube-comment-downloader
yt-dlp
vaderSentiment
google-generativeai
google-genai
geopy
folium
